finalized wordcloud
