+++ src/Components/Plots/WordCloud/Tooltip.tsx
@@ -1,0 +1,15 @@
+import * as d3 from "d3";
+
+export const createTooltip = () => {
+    return d3.select("body")
+        .append("div")
+        .attr("class", "tooltip")
+        .style("position", "absolute")
+        .style("background", "#f4f4f4")
+        .style("padding", "8px")
+        .style("border", "1px solid #ccc")
+        .style("border-radius", "4px")
+        .style("pointer-events", "none")
+        .style("opacity", 0)
+        .style("font", "var(--text-body-shorthand-medium)");
+};
